mirror-crawler/docker/requirements.txt
beautifulsoup4==4.8.2
cachetools==4.0.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-bigquery==1.24.0
google-cloud-container==0.4.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.27.2
idna==2.8
Jinja2==2.11.1
kubernetes==10.0.1
MarkupSafe==1.1.1
oauthlib==3.1.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-magic==0.4.15
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.8
websocket-client==0.57.0
